+++ src/backend/impress/settings.py
@@ -884,7 +884,12 @@ class Base(Configuration):
     USER_RECONCILIATION_FORM_URL = values.Value(
         None, environ_name="USER_RECONCILIATION_FORM_URL", environ_prefix=None
     )
-
+    USER_ONBOARDING_DOCUMENTS = values.ListValue(
+        [], environ_name="USER_ONBOARDING_DOCUMENTS", environ_prefix=None
+    )
+    USER_ONBOARDING_SANDBOX_DOCUMENT = values.Value(
+        None, environ_name="USER_ONBOARDING_SANDBOX_DOCUMENT", environ_prefix=None
+    )
     # Marketing and communication settings
     SIGNUP_NEW_USER_TO_MARKETING_EMAIL = values.BooleanValue(
         False,